Refs #32608 - Address comments
