+++ webpack/components/AnsibleHostDetail/components/AnsibleVariableOverrides/AnsibleVariableOverridesHelper.js
@@ -1,8 +1,8 @@
 export const reorderVariables = variables =>
   variables.reduce((memo, role) => {
-    const vars = role?.ansibleVariablesWithOverrides?.nodes.map(variable => ({
-      ...variable,
-      roleName: role.name,
-    }));
-    return memo.concat(vars);
+    const vars = role?.ansibleVariablesWithOverrides?.nodes;
+    if (vars) {
+      return memo.concat(vars);
+    }
+    return memo;
   }, []);
